Rename to "eawsudo" to avoid naming conflict
There is another product called "awsudo" which predates this one. Rename
to "easudo" -- it's like "awsudo" but starts with "ea".
