+++ bin/eawsudo
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # Copyright (C) 2015 Electronic Arts Inc.  All rights reserved.
 
-require 'awsudo'
+require 'eawsudo'
 
 def usage
   warn <<-EOS
@@ -14,7 +14,7 @@ end
 
 usage if ARGV.size < 2
 
-config = Hash[*File.read(File.join(ENV['HOME'], '.awsudo')).
+config = Hash[*File.read(File.join(ENV['HOME'], '.eawsudo')).
   scan(/^(\w+)\s*=\s*(.*)$/).flatten]
 AWSUDO.config(config['IDP_LOGIN_URL'], config['SAML_PROVIDER_NAME'])
