+++ lib/eawsudo.rb
